requirements.txt
# Core System Monitoring
psutil==5.9.8              # System and process utilities
watchdog==4.0.0            # File system monitoring

# Network Monitoring (optional - requires admin/root)
# scapy==2.5.0             # Network packet analysis (comment if permission issues)

# Machine Learning
scikit-learn==1.3.2        # ML algorithms (Isolation Forest, etc.)
numpy==1.24.3              # Numerical operations
pandas==2.0.3              # Data manipulation

# Database
# sqlite3 is built-in with Python

# Web Dashboard
Flask==3.0.0               # Web framework
Flask-CORS==4.0.0          # Cross-origin requests

# Visualization
matplotlib==3.8.2          # Plotting
plotly==5.18.0            # Interactive plots

# Logging
python-json-logger==2.0.7  # Structured JSON logging

# Configuration
PyYAML==6.0.1             # YAML config parsing

# Utilities
python-dateutil==2.8.2    # Date utilities
colorama==0.4.6           # Colored terminal output

# Testing (development)
pytest==7.4.3             # Testing framework
pytest-cov==4.1.0         # Coverage reports
